Move ArrowError to arrow-schema
